requirements.txt
azure-cosmos==4.3.0
pytest